Move InReplyToMetadata to own file
